+++ src/YTTrending.Application/Features/Videos/Queries/GetVideos/GetVideosQuery.cs
@@ -1,0 +1,5 @@
+using YTTrending.Application.Features.Videos.Dtos;
+
+namespace YTTrending.Application.Features.Videos.Queries.GetVideos;
+public record GetVideosQuery : VideoFilter, IRequest<Result<PagedResult<VideoDto>>>;
+